requirements.txt
selenium
pillow
webdriver-manager